% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_gaps}
\alias{check_gaps}
\title{Get Duality Gaps}
\usage{
check_gaps(fit, standardize, x, y, tol_gap = 1e-04)
}
\arguments{
\item{fit}{the resulting fit}

\item{standardize}{whether the fit used standardization}

\item{x}{design matrix}

\item{y}{response vector}

\item{tol_gap}{the tolerance threshold used when fitting}
}
\value{
primals, duals, and relative duality gaps
}
\description{
Get Duality Gaps
}
