% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HessianScreening-package.R
\docType{package}
\name{HessianScreening-package}
\alias{HessianScreening}
\alias{HessianScreening-package}
\title{HessianScreening: Hessian-Based Screening Rules for the Lasso}
\description{
Procedures to fit the lasso using Hessian-based screening rules in addition to several other screening rules.
}
\author{
\strong{Maintainer}: John Doe \email{johndoe@email.com}

}
\keyword{internal}
