% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lassoPath.R
\name{lassoPath}
\alias{lassoPath}
\title{Lasso Path with Hessian Screening Rules}
\usage{
lassoPath(
  X,
  y,
  family = c("gaussian", "binomial"),
  lambda = NULL,
  standardize = TRUE,
  screening_type = c("working", "hessian", "edpp", "gap_safe", "strong", "celer",
    "blitz", "sasvi"),
  shuffle = match.arg(screening_type) == "blitz",
  check_frequency = if (NROW(X) > NCOL(X)) 1 else 10,
  screen_frequency = 10,
  hessian_warm_starts = TRUE,
  celer_use_old_dual = TRUE,
  celer_use_accel = TRUE,
  celer_prune = TRUE,
  gap_safe_active_start = TRUE,
  augment_with_gap_safe = TRUE,
  log_hessian_update_type = c("full", "auto", "approx"),
  log_hessian_auto_update_freq = 10,
  path_length = 100L,
  maxit = 1e+05,
  tol_gap = 1e-04,
  gamma = 0.01,
  store_dual_variables = FALSE,
  verify_hessian = FALSE,
  line_search = TRUE,
  verbosity = 0
)
}
\arguments{
\item{X}{The predictor matrix}

\item{y}{The reponse vector}

\item{family}{The name of the family, "gaussian" or "logistic"}

\item{lambda}{weights for the regularization path, if \code{NULL}, then they
are automatically computed}

\item{standardize}{Whether to standardize the predictors}

\item{screening_type}{Screening rule}

\item{shuffle}{Shuffle working set before each CD pass?}

\item{check_frequency}{Frequency at which duality gap is checked for
inner CD loop}

\item{screen_frequency}{Frequency at which predictors are screened for
the gap safe solver}

\item{hessian_warm_starts}{Whether to use warm starts based on Hessian}

\item{celer_use_old_dual}{whether to try to use previous dual when checking
duality gap globally and screening for Celer}

\item{celer_use_accel}{whether to use accelerated dual point for Celer}

\item{celer_prune}{whether to use pruning for Celer}

\item{gap_safe_active_start}{Whether to use the active start strategy for
the Gap-Safe rule}

\item{augment_with_gap_safe}{Whether or not augment heuristic rules
with Gap Safe checks during KKT checks}

\item{log_hessian_update_type}{What type of strategy to use for
updating the hessian for logistic regression}

\item{log_hessian_auto_update_freq}{Frequency of hessian updates when
\code{log_hessian_update_type = "auto"}}

\item{path_length}{The (desired) length of the lasso path}

\item{maxit}{Maximum number of iterations for Coordinate Descent loop}

\item{tol_gap}{Tolerance threshold for relative duality gap.}

\item{gamma}{Percent of strong approximation to add to Hessian approximation}

\item{store_dual_variables}{whether to store dual variables throughout
fitting}

\item{verify_hessian}{Whether to not to verify that Hessian updates are
correct. Used only for diagnostic purposes.}

\item{line_search}{Use line search in CD solver.}

\item{verbosity}{Controls the level of verbosity. 0 = no output, 1 = outer
level output, 2 = inner solver output}
}
\description{
This function fits the full lasso path.
}
